refactor(model): 重构模型层代码
- 将 model 包拆分为 domain 和 dto 包- 将实体类统一放到 domain 包下
- 将请求和响应 DTO 类放到 dto 包下
- 更新相关引用和包名

+++ src/main/java/com/xc/blogbackend/controller/ResourceController.java
@@ -2,7 +2,7 @@ package com.xc.blogbackend.controller;
 
 import com.xc.blogbackend.common.BaseResponse;
 import com.xc.blogbackend.common.ResultUtils;
-import com.xc.blogbackend.model.domain.BgResource;
+import com.xc.blogbackend.model.domain.entity.BgResource;
 import com.xc.blogbackend.service.BgResourceService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -30,7 +30,7 @@ public class ResourceController {
      */
     @ApiOperation(value = "根据分类获取站点")
     @GetMapping("/sites/{categoryId}")
-    public BaseResponse<List<BgResource>> getSiteToCategory(@PathVariable Integer categoryId) {
+    public BaseResponse<List<BgResource>> getSiteToCategory(@PathVariable Long categoryId) {
         List<BgResource> bgResourceList = bgResourceService.getSiteToCategory(categoryId);
         return ResultUtils.success(bgResourceList);
     }
